+++ .github/workflows/gql-perf-arch-github_workflow.yml
@@ -2,7 +2,7 @@ name: gql-perf-arch-github_workflow
 on:
   push:
     branches:
-      - master
+      - main
   workflow_dispatch:
     inputs:
       COMMIT_HASH_INPUT:
@@ -25,18 +25,20 @@ env:
   COMMIT_HASH: ${{ (github.sha) }}
   CLASSES: ${{ (github.event.inputs.CLASSES_TO_EXECUTE_INPUT) }}
   PULL_REQUEST_NUMBER: ${{ (github.event.inputs.PULL_REQUEST_NUMBER_INPUT) }}
+
 jobs:
-  push_action:
+  execute_workflow:
     runs-on: ubuntu-latest
-    if: ${{ github.event_name == 'push' }}
     steps:
       - uses: actions/checkout@v3
       - uses: actions/setup-node@v3
         with:
           node-version: '14'
       - run: npm install
-      - run: npm i yenv
-      - run: npm i uuid
+
+      - name: Update COMMIT_HASH
+        if: ${{ github.event_name == 'workflow_dispatch' }}
+        run: echo "COMMIT_HASH=${{ (github.event.inputs.COMMIT_HASH_INPUT) }} " >> $GITHUB_ENV
 
       - id: 'auth'
         name: 'Authenticate to Google Cloud'
@@ -45,26 +47,4 @@ jobs:
           credentials_json: ${{ secrets.GOOGLE_APPLICATION_CREDENTIALS }}
 
       - name: Execute JS script
-        run: node google_cloud_http_target_task.js
-
-  workflow_dispatch_action:
-    runs-on: ubuntu-latest
-    if: ${{ github.event_name == 'workflow_dispatch' }}
-    steps:
-      - uses: actions/checkout@v3
-      - uses: actions/setup-node@v3
-        with:
-            node-version: '14'
-      - run: npm install
-      - run: npm i yenv
-      - run: npm install uuid
-      - run: echo "COMMIT_HASH=${{ (github.event.inputs.COMMIT_HASH_INPUT) }} " >> $GITHUB_ENV
-
-      - id: 'auth'
-        name: 'Authenticate to Google Cloud'
-        uses: google-github-actions/auth@v0.4.0
-        with:
-          credentials_json: ${{ secrets.GOOGLE_APPLICATION_CREDENTIALS }}
-
-      - name: Execute JS script
-        run: node google_cloud_http_target_task.js
+        run: node .github/workflows/google_cloud_http_target_task.js
