+++ package.json
@@ -13,11 +13,10 @@
   "dependencies": {
     "@google-cloud/tasks": "^3.0.0",
     "body-parser": "^1.18.3",
-    "express": "^4.16.3"
+    "express": "^4.16.3",
+    "yenv": "^3.0.1"
   },
   "devDependencies": {
-    "chai": "^4.2.0",
-    "mocha": "^8.0.0",
     "uuid": "^8.0.0"
   }
 }
